投稿の詳細ページの作成

+++ bbs_samples/view.php
@@ -1,3 +1,26 @@
+<?php
+session_start();
+require('library.php');
+
+if (isset($_SESSION['id']) && isset($_SESSION['name'])) {
+    $name = $_SESSION['name'];
+    $id = $_SESSION['id'];
+} else {
+    header('Location: login.php');
+    exit();
+}
+
+$id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_NUMBER_INPUT);
+if (!$id) {
+    header('Location : index.php');
+    exit();
+} 
+
+
+$db = dbconnect();
+
+
+?>
 <!DOCTYPE html>
 <html lang="ja">
 
@@ -17,14 +40,31 @@
     </div>
     <div id="content">
         <p>&laquo;<a href="index.php">一覧にもどる</a></p>
+        <?php $stmt = $db->prepare('select p.id p.member_id, p.message, p.created, m.name, m.picture from posts p, members m where p.id = ? and m.id = p.member_id by id desc');
+        if (!$stmt) {
+            die($db->error);
+        }
+        $stmt->bind_param('i', $id);
+        $success = $stmt->execute();
+        if (!$success) {
+            die($db->error);
+        }
+
+        $stmt->bind_result($id, $member_id, $message, $created, $name, $picture);
+        if($stmt->fetch()):
+        ?>
         <div class="msg">
-            <img src="member_picture/" width="48" height="48" alt=""/>
-            <p>○○<span class="name">（○○）</span></p>
-            <p class="day"><a href="view.php?id=">2021/01/01 00:00:00</a>
+            <?php if ($picture): ?>
+            <img src="member_picture/<?php echo h($picture) ?>" width="48" height="48" alt=""/>
+            <?php endif; ?>
+            <p><?php echo h($message); ?><span class="name">（<?php echo h($name); ?>）</span></p>
+            <p class="day"><a href="view.php?id="><?php echo h($created); ?></a>
                 [<a href="delete.php?id=" style="color: #F33;">削除</a>]
             </p>
         </div>
-        <p>その投稿は削除されたか、URLが間違えています</p>
+        <?php else: ?>
+            <p>その投稿は削除されたか、URLが間違えています</p>
+        <?php endif ;?>
     </div>
 </div>
 </body>
